🔨refactor: 테스트 목데이터 위치 변경
- TestFixtureFactory 위치 변경
